+++ tests/testthat/test-ordinal_reg-clm.R
@@ -1,5 +1,24 @@
 seed <- 144688L
 
+# specification: arguments -----------------------------------------------------
+
+test_that("specification handles model parameters", {
+  skip_if_not_installed("MASS")
+  house_sub <- get_house()$sub
+
+  # a legitimate ordinal link function not recognized by {dials}
+  ao_spec <- ordinal_reg(engine = "clm", ordinal_link = "Aranda-Ordaz")
+  expect_no_error(suppressWarnings(suppressMessages(
+    fit(ao_spec, Sat ~ Infl + Cont, data = house_sub)
+  )))
+
+  # an unavailable odds link function
+  acat_spec <- ordinal_reg(engine = "clm", odds_link = "adjacent_categories")
+  expect_snapshot(error = TRUE, {
+    fit(acat_spec, Sat ~ Type + Infl + Cont, data = house_sub)
+  })
+})
+
 # model: basic -----------------------------------------------------------------
 
 test_that("model object", {
@@ -29,13 +48,14 @@ test_that("model object", {
     Sat ~ Type + Infl + Cont,
     data = house_sub,
     link = "probit",
-    threshold = "equidistant"
+    threshold = "symmetric"
   )
 
-  tidy_spec <- ordinal_reg() |>
+  tidy_spec <- ordinal_reg(
+    ordinal_link = "probit", threshold_structure = "symmetric_median"
+  ) |>
     set_engine("clm") |>
-    set_mode("classification") |>
-    set_args(ordinal_link = "probit", threshold = "equidistant")
+    set_mode("classification")
   set.seed(seed)
   tidy_fit <- fit(tidy_spec, Sat ~ Type + Infl + Cont, data = house_sub)
 
@@ -186,4 +206,64 @@ test_that("arguments agree", {
     )
   })
   expect_equal(clm_arg_fit$fit$link, "probit")
+})
+
+# parallel regression ----------------------------------------------------------
+
+test_that("parallel regression argument handles logicals", {
+  skip_if_not_installed("MASS")
+  skip_if_not_installed("ordinal")
+  house_sub <- get_house()$sub
+
+  # all parallel regression
+
+  set.seed(seed)
+  tidy_fit <- ordinal_reg(parallel_reg = TRUE, engine = "clm") |>
+    fit(Sat ~ Infl + Cont, data = house_sub)
+
+  set.seed(seed)
+  orig_fit <- ordinal::clm(
+    Sat ~ Infl + Cont, data = house_sub
+  )
+
+  tidy_fit$fit$call <- orig_fit$call <- NULL
+  tidy_fit$fit$formulas <- orig_fit$formulas <- NULL
+  expect_equal(tidy_fit$fit, orig_fit, ignore_formula_env = TRUE)
+
+  # all category-specific
+
+  set.seed(seed)
+  tidy_fit <- ordinal_reg(parallel_reg = FALSE, engine = "clm") |>
+    fit(Sat ~ Infl + Cont, data = house_sub)
+
+  set.seed(seed)
+  orig_fit <- ordinal::clm(
+    Sat ~ 1, data = house_sub,
+    nominal = ~ Infl + Cont
+  )
+
+  tidy_fit$fit$call <- orig_fit$call <- NULL
+  tidy_fit$fit$formulas <- orig_fit$formulas <- NULL
+  expect_equal(tidy_fit$fit, orig_fit, ignore_formula_env = TRUE)
+
+  # all category-specific with interactions
+
+  set.seed(seed)
+  suppressWarnings(
+    tidy_fit <- ordinal_reg(parallel_reg = FALSE, engine = "clm") |>
+      fit(Sat ~ Infl * Cont, data = house_sub)
+  )
+
+  set.seed(seed)
+  suppressWarnings(
+    orig_fit <- ordinal::clm(
+      Sat ~ 1, data = house_sub,
+      # wrapper expands full interaction expressions into summands
+      nominal = ~ Infl + Cont + Infl:Cont
+    )
+  )
+
+  tidy_fit$fit$call <- orig_fit$call <- NULL
+  tidy_fit$fit$formulas <- orig_fit$formulas <- NULL
+  expect_equal(tidy_fit$fit, orig_fit, ignore_formula_env = TRUE)
 })